Add files via upload
Develop a web application where a Servlet determine a person’s eligibility for
voting, based on his/her age. A HTML form receives the person’s name and age,
and supplies to Servlet for processing. If person’s age is &gt;=18 years, then display
the message as “Hello personName, you are eligible for voting”. Otherwise, display
the message as “Hello personName, youare not eligible for voting”. The
forms/output should appear as follows.