Add files via upload
Create an Html page that contains 4 option buttons Java, UNIX, DDBMS, OOSE
and 2 buttons Submit and Reset. When the user clicks on Submit button the server
responds by adding cookie containing the selected Subject and sends the html page
to the client. Program should not allow duplicate cookie to be written.