Add files via upload
Develop a web application where a HTML form allows the user to register with
following parameters (name, age, phone, email). Store all parameters’ values in
cookies and disply them on a Servlet for 2 minutes only. After displaying the
information, the Servlet redirects the request from one host to another host

(www.google.in), giving an explanation to the client before redirection. In online
shopping, such situations are common where we are redirected from e-commerce
portal to bank for payment.